fix(ui): bottom-sheet bg gradient banding
